+++ tools/addstring.sh
@@ -1,12 +1,10 @@
 #!/bin/bash
-grep -RIl "\<string name=\"help_check_restrict" res | xargs sed -i -e '/help_check_restrict/d'
-grep -RIl "\<string name=\"help_check_ondemand" res | xargs sed -i -e '/help_check_ondemand/d'
-grep -RIl "\<string name=\"help_switch" res | xargs sed -i -e '/help_switch/d'
+grep -RIl "\<string name=\"title_template_merge_reset" res | xargs sed -i -e '/title_template_merge_reset/d'
 
-grep -RIl "\<string name=\"settings_aosp" res | xargs sed -i -e '/settings_aosp/a \
-\ \ \ \ <string name=\"help_check_restrict\">First check box: restrict category or function</string>\
-\ \ \ \ <string name=\"help_check_ondemand\">Second check box: restrict on demand</string>\
-\ \ \ \ <string name=\"help_switch\">On/off switch: enable or disable all restrictions</string>'
+grep -RIl "\<string name=\"title_template_merge" res | xargs sed -i -e '/title_template_merge/a \
+\ \ \ \ <string name=\"title_template_merge_reset\">Apply template (merge reset)</string>'
+
+grep -RIl "\<string name=\"title_template_merge" res | xargs sed -i -e 's/Apply template (merge)/Apply template (merge set)/g'
 
 #grep -RIl "\<string name=\"app_notxposed" res | xargs sed -i -e '/app_notxposed/d'
 #grep -RIl "\<string name=\"restrict_help_internet" res | xargs sed -i -e 's/internet</Internet</g'